A basic Vue.js/Bootstrap app for editing Phyloreference testcases (#5)
Created an index.html file and curation-tool.js script that displays phyloreference test cases in a Vue.js application. It can load JSON files from the user's computer or one of the included example files, and display included phylogenies (using Phylotree.js) and phyloreferences (as descriptions and a list of specifiers).