Create Github Action for deploying to Github Pages
I've based this on an existing workflow in the Github Marketplace. It should be triggered every time we publish a new release. Once triggered, it checks out that release, builds the HTML and JavaScript using Vue, and deploys it to the gh-pages branch using an SSH deploy key stored as a secret.